requirements.txt
biopython==1.70
pandas==0.19.0
pysam==0.14